# commerceue-app/WebContent/swagger_v3.0.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"API Extensions for IBM Digital Commerce customizations.",
      "version":"3.0",
      "title":"IBM Digital Commerce API Extensions"
   },
   "basePath":"/commerceue/extension",
...
   "paths":{  
...
      "/promotion/process_promotion_code":{  
         "post":{  
            "tags":[  
               "promotion"
            ],
            "summary":"The process_promotion_code command processes the promotion code - return the real promotion code, or throw exception.",
            "description":"| Command: com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd |\n|----------|\n| PromotionCodeAddRemoveControllerCmd is a controller command which is used to add or remove a promotion code from an order.|",
            "operationId":"processPromotionCode",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"processPromotionCode UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/OrderPreUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/OrderPreUEResponse"
                  }
               },
               "400":{  
                  "description":"Invalid Input Parameter"
               },
               "500":{  
                  "description":"Internal server error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd"
            }
         }
      },
      "/promotion/promotion_code_add_remove_post":{  
         "post":{  
            "tags":[  
               "promotion"
            ],
            "summary":"Customize actions to perform after adding/removing promotion codes to/from the cart.",
            "description":"| Command: com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd |\n|----------|\n| PromotionCodeAddRemoveControllerCmd is a controller command which is used to add or remove a promotion code from an order.|",
            "operationId":"promotionCodeAddRemovePost",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"promotionCodeAddRemovePost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/OrderPostUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/OrderPreUEResponse"
                  }
               },
               "400":{  
                  "description":"Invalid Input Parameter"
               },
               "500":{  
                  "description":"Internal server error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd"
            }
         }
      },
      "/promotion/promotion_code_add_remove_pre":{  
         "post":{  
            "tags":[  
               "promotion"
            ],
            "summary":"Customize actions to perform before adding/removing promotion codes to/from the cart.",
            "description":"| Command: com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd |\n|----------|\n| PromotionCodeAddRemoveControllerCmd is a controller command which is used to add or remove a promotion code from an order.|",
            "operationId":"promotionCodeAddRemovePre",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"promotionCodeAddRemovePre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/OrderPreUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/OrderPreUEResponse"
                  }
               },
               "400":{  
                  "description":"Invalid Input Parameter"
               },
               "500":{  
                  "description":"Internal server error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.marketing.commands.PromotionCodeAddRemoveControllerCmd"
            }
         }
      },
...
   },
   "definitions":{  
...
      "AdjustmentRequirement":{  
         "type":"object",
         "properties":{  
            "adjustmentUsage":{  
               "type":"string"
            },
            "adjustmentCategory":{  
               "type":"string"
            },
            "adjustmentValue":{  
               "type":"string"
            },
            "comments":{  
               "$ref":"#/definitions/Description"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "ContactInfo":{  
         "type":"object",
         "properties":{  
            "contactInfoIdentifier":{  
               "$ref":"#/definitions/ContactInfoIdentifier"
            },
            "contactName":{  
               "$ref":"#/definitions/PersonName"
            },
            "address":{  
               "$ref":"#/definitions/Address"
            },
            "telephone1":{  
               "$ref":"#/definitions/Telephone"
            },
            "telephone2":{  
               "$ref":"#/definitions/Telephone"
            },
            "bestCallingTime":{  
               "type":"string"
            },
            "emailAddress1":{  
               "$ref":"#/definitions/EmailAddress"
            },
            "emailAddress2":{  
               "$ref":"#/definitions/EmailAddress"
            },
            "fax1":{  
               "$ref":"#/definitions/Fax"
            },
            "fax2":{  
               "$ref":"#/definitions/Fax"
            },
            "mobilePhone1":{  
               "$ref":"#/definitions/MobilePhone"
            },
            "organizationName":{  
               "type":"string"
            },
            "organizationUnitName":{  
               "type":"string"
            },
            "geographicalShippingCode":{  
               "type":"string"
            },
            "geographicalTaxCode":{  
               "type":"string"
            },
            "attributes":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "language":{  
               "type":"string"
            }
         }
      },
...
      "PromotionCode":{  
         "type":"object",
         "properties":{  
            "code":{  
               "type":"string"
            },
            "reason":{  
               "$ref":"#/definitions/PromotionCodeReason"
            },
            "associatedPromotion":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/AssociatedPromotion"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "RewardOption":{  
         "type":"object",
         "properties":{  
            "rewardOptionIdentifier":{  
               "$ref":"#/definitions/RewardOptionIdentifier"
            },
            "adjustment":{  
               "$ref":"#/definitions/Adjustment"
            },
            "rewardChoice":{  
               "$ref":"#/definitions/RewardChoice"
            },
            "rewardSpecification":{  
               "$ref":"#/definitions/RewardSpecification"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "PaymentInformation":{  
         "type":"object",
         "properties":{  
            "usableBillingAddress":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ContactInfoIdentifier"
               }
            },
            "paymentMethod":{  
               "$ref":"#/definitions/PaymentMethod"
            },
            "paymentTermConditionId":{  
               "type":"string"
            },
            "protocolData":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ProtocolData"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "OrderShippingInfo":{  
         "type":"object",
         "properties":{  
            "shippingAddress":{  
               "$ref":"#/definitions/ContactInfo"
            },
            "shippingMode":{  
               "$ref":"#/definitions/ShippingMode"
            },
            "requestedShipDate":{  
               "type":"string",
               "format":"date-time"
            },
            "physicalStoreIdentifier":{  
               "$ref":"#/definitions/PhysicalStoreIdentifier"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "OrderStatus":{  
         "type":"object",
         "properties":{  
            "status":{  
               "type":"string"
            },
            "buyerApprovalStatus":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "prepareIndicator":{  
               "type":"boolean",
               "default":false
            },
            "editable":{  
               "type":"boolean",
               "default":false
            },
            "locked":{  
               "type":"boolean",
               "default":false
            },
            "blocked":{  
               "type":"boolean",
               "default":false
            }
         }
      },
...
      "OrderPostUERequest":{  
         "type":"object",
         "properties":{  
            "contextData":{  
               "$ref":"#/definitions/ContextData"
            },
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "commandOutputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "order":{  
               "$ref":"#/definitions/Order"
            }
         }
      },
...
      "StoreIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "externalIdentifier":{  
               "$ref":"#/definitions/StoreExternalIdentifier"
            }
         }
      },
...
      "CouponCode":{  
         "type":"object",
         "properties":{  
            "couponId":{  
               "type":"string"
            },
            "code":{  
               "type":"string"
            },
            "promotionIdentifier":{  
               "$ref":"#/definitions/PromotionIdentifier"
            },
            "description":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/CouponDescription"
               }
            },
            "effectiveDateTime":{  
               "type":"string",
               "format":"date-time"
            },
            "expirationDateTime":{  
               "type":"string",
               "format":"date-time"
            },
            "status":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "OrderPaymentInfo":{  
         "type":"object",
         "properties":{  
            "paymentInstruction":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/PaymentInstruction"
               }
            },
            "status":{  
               "type":"string"
            },
            "financialTransaction":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/FinancialTransaction"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "Locale":{  
         "type":"object",
         "properties":{  
            "language":{  
               "type":"string"
            },
            "script":{  
               "type":"string"
            },
            "country":{  
               "type":"string"
            },
            "variant":{  
               "type":"string"
            },
            "extensionKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleAttributes":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "iso3Language":{  
               "type":"string"
            },
            "iso3Country":{  
               "type":"string"
            },
            "displayLanguage":{  
               "type":"string"
            },
            "displayScript":{  
               "type":"string"
            },
            "displayCountry":{  
               "type":"string"
            },
            "displayVariant":{  
               "type":"string"
            },
            "displayName":{  
               "type":"string"
            }
         }
      },
...
      "OrderIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "externalOrderID":{  
               "type":"string"
            },
            "customerOrderNumber":{  
               "type":"string"
            }
         }
      },
...
      "QuoteIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "externalQuoteID":{  
               "type":"string"
            }
         }
      },
...
      "PersonIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "distinguishedName":{  
               "type":"string"
            },
            "externalIdentifier":{  
               "$ref":"#/definitions/PersonExternalIdentifier"
            }
         }
      },
...
      "ShippingMode":{  
         "type":"object",
         "properties":{  
            "shippingModeIdentifier":{  
               "$ref":"#/definitions/ShippingModeIdentifier"
            },
            "description":{  
               "$ref":"#/definitions/Description"
            },
            "trackingURL":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
      "ContextData":{  
         "type":"object",
         "properties":{  
            "baseContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "accessControlContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "locale":{  
               "$ref":"#/definitions/Locale"
            }
         }
      },
      "ContactInfoIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "externalIdentifier":{  
               "$ref":"#/definitions/ContactInfoExternalIdentifier"
            }
         }
      },
...
      "Order":{  
         "type":"object",
         "properties":{  
            "orderIdentifier":{  
               "$ref":"#/definitions/OrderIdentifier"
            },
            "storeIdentifier":{  
               "$ref":"#/definitions/StoreIdentifier"
            },
            "channel":{  
               "$ref":"#/definitions/Channel"
            },
            "quoteID":{  
               "type":"string"
            },
            "quoteIdentifier":{  
               "$ref":"#/definitions/QuoteIdentifier"
            },
            "buyerPONumber":{  
               "type":"string"
            },
            "organizationIdentifier":{  
               "$ref":"#/definitions/OrganizationIdentifier"
            },
            "buyerIdentifier":{  
               "$ref":"#/definitions/PersonIdentifier"
            },
            "buyerBillInfo":{  
               "$ref":"#/definitions/ContactInfo"
            },
            "couponCode":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/CouponCode"
               }
            },
            "promotionCode":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/PromotionCode"
               }
            },
            "rewardOption":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/RewardOption"
               }
            },
            "adjustmentRequirement":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/AdjustmentRequirement"
               }
            },
            "orderAmount":{  
               "$ref":"#/definitions/OrderCharges"
            },
            "usableShippingAddress":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ContactInfoIdentifier"
               }
            },
            "orderPaymentInfo":{  
               "$ref":"#/definitions/OrderPaymentInfo"
            },
            "usablePaymentInformation":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/PaymentInformation"
               }
            },
            "usableShippingMode":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ShippingMode"
               }
            },
            "orderStatus":{  
               "$ref":"#/definitions/OrderStatus"
            },
            "orderEditor":{  
               "$ref":"#/definitions/PersonIdentifier"
            },
            "orderDescription":{  
               "type":"string"
            },
            "comments":{  
               "type":"string"
            },
            "lastUpdateDate":{  
               "type":"string",
               "format":"date-time"
            },
            "placedDate":{  
               "type":"string",
               "format":"date-time"
            },
            "orderItem":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/OrderItem"
               }
            },
            "orderNotificationInfo":{  
               "$ref":"#/definitions/OrderNotificationInfo"
            },
            "orderShippingInfo":{  
               "$ref":"#/definitions/OrderShippingInfo"
            },
            "orderScheduleInfo":{  
               "$ref":"#/definitions/OrderScheduleInfo"
            },
            "orderGiftInfo":{  
               "$ref":"#/definitions/GiftItemInfo"
            },
            "parentInfo":{  
               "$ref":"#/definitions/ParentInfo"
            },
            "subscriptionInfo":{  
               "$ref":"#/definitions/SubscriptionInfo"
            },
            "calculationInfo":{  
               "$ref":"#/definitions/CalculationInfo"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "orderExtendAttribute":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ExtendAttribute"
               }
            },
            "orderVersion":{  
               "type":"string"
            },
            "shipAsComplete":{  
               "type":"boolean",
               "default":false
            },
            "blockInfo":{  
               "$ref":"#/definitions/OrderBLockInfo"
            },
            "csrcomments":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/CSRComments"
               }
            },
            "orderCode":{  
               "type":"string"
            }
         }
      },
...
      "Channel":{  
         "type":"object",
         "properties":{  
            "channelIdentifer":{  
               "$ref":"#/definitions/ChannelIdentifer"
            },
            "description":{  
               "$ref":"#/definitions/Description"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
      "PersonExternalIdentifier":{  
         "type":"object",
         "properties":{  
            "identifier":{  
               "type":"string"
            }
         }
      },
      "OrganizationIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "distinguishedName":{  
               "type":"string"
            }
         }
      },
...
      "OrderScheduleInfo":{  
         "type":"object",
         "properties":{  
            "startTime":{  
               "type":"string",
               "format":"date-time"
            },
            "interval":{  
               "type":"integer",
               "format":"int32"
            }
         }
      },
...
      "OrderCharges":{  
         "type":"object",
         "properties":{  
            "grandTotal":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "totalProductPrice":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "totalAdjustment":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "genericTotal":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/GenericTotal"
               }
            },
            "adjustment":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/Adjustment"
               }
            },
            "totalShippingCharge":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "totalSalesTax":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "totalShippingTax":{  
               "$ref":"#/definitions/MonetaryAmount"
            },
            "totalTaxByTaxCategory":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/TaxByTaxCategory"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "OrderPreUERequest":{  
         "type":"object",
         "properties":{  
            "contextData":{  
               "$ref":"#/definitions/ContextData"
            },
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "order":{  
               "$ref":"#/definitions/Order"
            }
         }
      },
      "OrderPreUEResponse":{  
         "type":"object",
         "properties":{  
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "order":{  
               "description":"The order pojo object",
               "$ref":"#/definitions/Order"
            }
         },
         "description":"This is the pre UE response for the order controller commands"
      },
...
   }
}